refactor: Move run-deleting code into a separate class
