+++ modules/extensions/galasa-extensions-parent/dev.galasa.ras.couchdb/src/main/java/dev/galasa/ras/couchdb/internal/CouchdbDirectoryService.java
@@ -15,8 +15,6 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map.Entry;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionException;
 import java.util.Set;
 
 import javax.validation.constraints.NotNull;
@@ -55,7 +53,6 @@ import dev.galasa.framework.spi.ras.ResultArchiveStoreFileStore;
 import dev.galasa.framework.spi.utils.GalasaGson;
 import dev.galasa.extensions.common.api.LogFactory;
 import dev.galasa.extensions.common.couchdb.CouchdbException;
-import dev.galasa.extensions.common.couchdb.pojos.IdRev;
 import dev.galasa.extensions.common.couchdb.pojos.Row;
 import dev.galasa.extensions.common.couchdb.pojos.ViewResponse;
 import dev.galasa.extensions.common.couchdb.pojos.ViewRow;
@@ -465,62 +462,6 @@ public class CouchdbDirectoryService implements IResultArchiveStoreDirectoryServ
         }
 
         return runs;
-    }
-
-    public void discardRun(@NotNull TestStructureCouchdb runTestStructure) throws ResultArchiveStoreException {
-        try {
-            // Build a list of discard operation futures
-            List<CompletableFuture<Void>> futures = discardRecords(LOG_DB, runTestStructure.getLogRecordIds());
-            futures.addAll(discardRecords(ARTIFACTS_DB, runTestStructure.getArtifactRecordIds()));
-            futures.add(discardRecord(RUNS_DB, runTestStructure._id, runTestStructure._rev));
-
-            // Wait for all the discard operations to finish
-            CompletableFuture.allOf(futures.toArray(new CompletableFuture[0])).join();
-        } catch (CompletionException e) {
-            throw new ResultArchiveStoreException("Failed to discard run: " + runTestStructure._id, e);
-        }
-    }
-
-    private List<CompletableFuture<Void>> discardRecords(String databaseName, List<String> ids) throws ResultArchiveStoreException {
-        List<CompletableFuture<Void>> futures = new ArrayList<>();
-        for (String id : ids) {
-            futures.add(discardRecord(databaseName, id));
-        }
-        return futures;
-    }
-
-    private CompletableFuture<Void> discardRecord(String databaseName, String id) throws ResultArchiveStoreException {
-        return getRevision(databaseName, id)
-            .thenCompose(revision -> discardRecord(databaseName, id, revision));
-    }
-
-    private CompletableFuture<Void> discardRecord(String databaseName, String id, String revision) {
-        return CompletableFuture.runAsync(() -> {
-            try {
-                store.deleteDocumentFromDatabase(databaseName, id, revision);
-            } catch (CouchdbException e) {
-                throw new CompletionException(e);
-            }
-        });
-    }
-
-    private CompletableFuture<String> getRevision(String databaseName, String id) {
-        return CompletableFuture.supplyAsync(() -> {
-            String foundRevision = null;
-            try {
-                IdRev found = store.getDocumentFromDatabase(databaseName, id, IdRev.class);
-                if (found._id == null) {
-                    throw new CouchdbRasException("Unable to find runs - Invalid JSON response");
-                }
-                if (found._rev == null) {
-                    throw new CouchdbRasException("Unable to find rev - Invalid JSON response");
-                }
-                foundRevision = found._rev;
-            } catch (CouchdbException | ResultArchiveStoreException e) {
-                throw new CompletionException(e);
-            }
-            return foundRevision;
-        });
     }
 
     private JsonObject buildGetRunsQuery(IRasSearchCriteria... searchCriterias) throws ResultArchiveStoreException {
