+++ modules/extensions/galasa-extensions-parent/dev.galasa.ras.couchdb/src/main/java/dev/galasa/ras/couchdb/internal/CouchdbRunResult.java
@@ -12,6 +12,7 @@ import dev.galasa.framework.spi.IRunResult;
 import dev.galasa.framework.spi.ResultArchiveStoreException;
 import dev.galasa.framework.spi.ras.ResultArchiveStoreFileStore;
 import dev.galasa.framework.spi.teststructure.TestStructure;
+import dev.galasa.ras.couchdb.internal.operations.CouchdbDeleteRunService;
 import dev.galasa.ras.couchdb.internal.pojos.TestStructureCouchdb;
 
 public class CouchdbRunResult implements IRunResult {
@@ -19,10 +20,12 @@ public class CouchdbRunResult implements IRunResult {
     private final TestStructureCouchdb   testStructure;
     private final CouchdbRasStore store;
     private final CouchdbDirectoryService storeService;
+    private final CouchdbDeleteRunService deleteRunService;
     private Path path;
 
     public CouchdbRunResult(CouchdbRasStore store, TestStructureCouchdb testStructure, LogFactory logFactory) {
         this.store = store;
+        this.deleteRunService = new CouchdbDeleteRunService(this.store);
         this.storeService = (CouchdbDirectoryService) store.getDirectoryServices().get(0);
         if (testStructure == null) {
             this.testStructure = new TestStructureCouchdb();
@@ -52,7 +55,7 @@ public class CouchdbRunResult implements IRunResult {
 
 	@Override
 	public void discard() throws ResultArchiveStoreException {
-        storeService.discardRun(this.testStructure);
+        deleteRunService.discardRun(this.testStructure);
 	}
 
     @Override
