+++ modules/extensions/galasa-extensions-parent/settings.gradle
@@ -1,5 +1,6 @@
 pluginManagement {
     repositories {
+        mavenLocal()
         maven {
             url = "$sourceMaven"
         }
